Refactor the whole project such that all apps live in one apps module.
