+++ apps/app_groceries/mapping.json
