+++ apps/app_groceries/pluralization_replacements.json
