+++ apps/app_groceries/setups.json
