+++ apps/app_wordhunt/pure_wordhunt.py
